feat: update version code to 18 and version name to 1.0.18; refactor data types for carat, taxRate, and exchangeRate to use kotlin.Double

+++ android/app/src/main/java/id/my/rizalanggoro/arta/feature/gold/presentation/tax/ListGoldTaxScreen.kt
@@ -40,7 +40,6 @@ import id.my.rizalanggoro.arta.shared.component.ConfirmDialog
 import id.my.rizalanggoro.arta.shared.component.EmptyPlaceholder
 import id.my.rizalanggoro.arta.shared.component.ErrorPlaceholder
 import id.my.rizalanggoro.arta.ui.theme.ArtaTheme
-import java.math.BigDecimal
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
@@ -226,16 +225,16 @@ private fun Preview() {
                     DtoGoldTaxPreference(
                         id = 1,
                         userId = 1,
-                        carat = BigDecimal.valueOf(24.0),
-                        taxRate = BigDecimal.valueOf(5.0),
+                        carat = 24.0,
+                        taxRate = 5.0,
                         createdAt = "",
                         updatedAt = ""
                     ),
                     DtoGoldTaxPreference(
                         id = 2,
                         userId = 1,
-                        carat = BigDecimal.valueOf(18.0),
-                        taxRate = BigDecimal.valueOf(3.5),
+                        carat = 18.0,
+                        taxRate = 3.5,
                         createdAt = "",
                         updatedAt = ""
                     ),
